Fix tbtc signer material symbols for non-frost builds
