+++ pkg/frost/signing/native_tbtc_signer_engine_frost_native.go
@@ -4,23 +4,6 @@ package signing
 
 import "fmt"
 
-const (
-	// NativeSignerMaterialFormatFrostTBTCSignerV1 carries signer material for
-	// tbtc-signer coarse session APIs.
-	NativeSignerMaterialFormatFrostTBTCSignerV1 = "frost-tbtc-signer-v1"
-	// NativeTBTCSignerKeyGroupSourceLegacyWalletPubKey marks scaffold-era
-	// key-group derivation from the legacy wallet public key.
-	NativeTBTCSignerKeyGroupSourceLegacyWalletPubKey = "legacy-wallet-pubkey"
-)
-
-// NativeTBTCSignerMaterialPayload is the signer-material payload schema for
-// `frost-tbtc-signer-v1`.
-type NativeTBTCSignerMaterialPayload struct {
-	KeyGroup                 string `json:"keyGroup"`
-	KeyGroupSource           string `json:"keyGroupSource,omitempty"`
-	LegacyPrivateKeyShareHex string `json:"legacyPrivateKeyShareHex,omitempty"`
-}
-
 // NativeTBTCSignerDKGParticipant identifies a DKG participant for coarse
 // tbtc-signer RunDKG operation.
 type NativeTBTCSignerDKGParticipant struct {
